requirements.txt
django==4.0.5
django-cors-headers==3.13.0
djangorestframework==3.13.1
django-environ==0.9.0
pillow==9.2.0